fix(btrfs-fs): Hardcode uids to be root
Add a patch to btrfs-progs that hardcodes all uid/gid in resulting images to be 0.
The default behavior is to inherit uid/gid from the the spcified rootDir.
This means that the resulting image ends up with /nix and /boot owned by the
uid/gid of the building user (usually 1000). Upon attempting to then boot this
image systemd-tmpfiles fails in strange ways, like failing to link /run -> /var/run
and failing to create /var/empty, degrading some system services that rely
on those (sshd and nscd).